rename EndpointTests to make way for a file that actually unit tests Endpoint methods
